homeworks/customerSegmentationWithDBSCAN/requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9 